Allow walk-in candidates to sign up to closed events
If an event is closed we usually prevent the candidate for signing up. If the
candidate visits the 'walk-in' URL of the event, we want to let them sign up.

Check for a `?walk_in` query parameter and pre-populate the store if present.

Allow candidate to go through the sign up journey if the event is closed and
they are a walk-in (otherwise keep the current behaviour of redirecting back to
the event page).