Add some CSS to better-style the blog quotes
The styles are more-or-less lifted from the Welcome Guide, but here we
have less control over the markup (as it's plain markdown) so the CSS is
a bit "smarter" with regards to not quoting (and making smaller) the
final lines of multi-paragraph quotes.

As in the comment below, we can revisit and improve this if it becomes a
hindrance, but for the time being living by the rule of "always cite
multiline quotes" should be enough!