Make body element display as a flex column
We currently have a bug where the <main> element being short means that
the footer isn't at the bottom of the screen. This looks bad but is
admittedly a minor problem.

The simple fix is to make <body> flex and allow <main> to grow. This
will push the footer to the bottom of the viewport.

https://developer.mozilla.org/en-US/docs/Web/CSS/flex-grow