Remove breadcrumbs from main content pages
It doesn't really make sense to include a breadcrumbs list with one
entry ('Home') on pages that are siblings to the home page. Home is also
represented in the main nav and is accessible by clicking the logo.

We should only show breadcrumbs on pages that are more than one level
deep.