Remove query string from script_src CSP directive
Some browsers display the following error when the query string part of
the directive is present.

> The source list for Content Security Policy directive 'script-src'
> contains a source with an invalid path: '/gtm.js?id=GTM-K7TK4WM'. The
> query component, including the '?', will be ignored.

It appears[0] that the query string isn't expected here and there's no
reference to it anywhere in the standard, the question mark isn't a
valid character so we can assume that part of the URL isn't expected.

[0] https://www.w3.org/TR/CSP3/#framework-directive-source-list