Anchor pagination links to the top of current list
Currently we use a JavaScript plugin to _remember_ how far we've
scrolled, so the pagination isn't jarring (#969), but this can be
unpredictable - if one page has fewer events on it, the user might end
up looking at a different category or at another part of the page
(usually the footer).

This is an attempt to improve that behaviour by adding anchors to the
events list(s) and adding an `anchor` param to the paginator. Now, on
every page change, the browser will anchor itself to the top of the
current list.