Favour heading over title for navigation text
This gives us an extra level of flexibility. The navigation_title takes
precendence, followed by the heading and if neither of those are set we
fall back to title.

Some pages require different things to be set; for example the home
page's title is currently "Inspire the next generation" but in the nav
it's called "Home". Other pages have a different title and heading (e.g.
"Teacher salaries and benefits" vs "Salaries and benefits".