Add AspectRatioController
We need to be able to set the aspect ratio for YouTube videos in order to
maintain a full preview thumbnail (without it being cropped).

We can do this with some CSS by wrapping the `iframe` and applying an
appropriate `padding-bottom` to the wrapper.

It would have been possible to do this just in CSS by manually working out the
aspect ratios for the videos and setting the appropriate padding, however I
figured it would be more user friendly to be able to take the width/height and
have a Stimulus controller do it for you (initially I thought the videos would
be a mixture of aspect ratios but it turned out they're all the same...). The
downside to this is the few with JS disabled will still get cropped previews.