Add ArrowComponent
We need to be able to display arrows on the welcome guide between different
bits of content.

Instead of having multiple SVGs for the different arrows pointing in various
directions and with various amount of curving, this adds an `ArrowComponent`
that - given points on a bezier path and an arrow head size - will render the
arrow dynamically.

We can then color and position the arrow in CSS, which should make it easier to
maintain and update in the future.

Includes previews for various arrow orientations.