Include Google structured data for how-to
Adding structured data for 'how-tos' enables Google to enhance the search
results for certain pages, showing the steps in the results.

Initially I've added this to the Steps to Become a Teacher page with mostly
dummy data; it'll need running past a content designer and designer for the
text/images, but its behind a feature flag just now.