Improve tab ordering in the cookie popup
The previous implementation had a couple of slight usability problems
that this (hopefully) addresses.

One was that the 'cookies' link (to our cookies page) wasn't accessible
via keyboard, tabbing only went between the acccpt/reject buttons.

The initial attempt to restrict tabbing to just one part of the page
(the modal) failed, even when using the new Dialog API - as did adding
`inert` in various places to prevent tabbing fron continuing to the
obstructed background.

Additionally, any click to the modal would move focus onto the next
element. This could have unintentional consequences for users who
misclick.

The approach I took was to just add a third target to the component and
set up the tab order manually. It works for both <tab> and <shift-tab>.

The event was changed from `blur` to `keydown` to stop clicking and
other interactions moving the focus along unintentionally.

Testing tabbing also appears to be painful[0] without bringing something
that actually drives the browser along to the party.

[0] https://stackoverflow.com/questions/1601593/fire-tab-keypress-event-in-javascript