Put each structured data section behind feature flag
Some of the structured data is fine to be shipped just now,
but other sections (such as events and blogs) need more
discussion outwith our team to get a green light.

Wrap each section behind a feature flag and enable the
breadcrumb list in production; we're going to use this one
to ensure its working and determine how long it would take
for Google to realise the change (so we get an idea of how
long a 'revert' would take if we make a mistake elsewhere,
as there's no way to preview what the search results will
look like ahead of making them live).