Move chevron into CSS
The `git-link` hyperlinks have a chevron at the end that is currently a
`<span>` with font awesome classes applied. This removes the need for the
`<span>` and instead specifies the icon in the `.git-link` CSS class, which
means we can use this type of link via the `chat_link` helper more easily.