Add fallback support to chat online buttons
We reference the chat online buttons in a number of different places, namely:

- In the 'talk to us' section of the footer
- As a card on the funding page
- As a CTA on the steps to become a teacher page
- In a explore section component on the individual story pages

When JS is enabled these all work fine, however we need to provide fallback
behaviour for when JS is disabled. For most cases we can change the button text
slightly and have it take them down the page to the 'talk to us' section, which
provides a telephone number and TTA CTA. In other cases, such as within the
talk to us section itself, we can simply hide the button.