Fix issue with Zendesk Chat and Turbolinks
The Zendesk Chat snippet appears to have logic in it to prevent the widget
being accidently loaded on the page multiple times. This prevents it from
re-loading on a page change by Turbolinks even when included in the `<body>`
tag. To get around this we need to clear `window.zEACLoaded` so that the script
re-evaluates.