Handle when the chat feature is offline
When the chat is unavailable we want to have 'Chat closed.' in the talk to us
section. In other places where a chat button is shown we either want to hide
the chat button completely (if not referenced by surrounding copy) or replace
it with some offline text to inform the user that its unavailable at the moment
but will be online between certain days/times.