Use JS to handle when chat is offline
The first attempt at this was in Rails and had an issue due to the way we cache
the static pages (using etags which change on deployment). To get around the
issue where someone sees the incorrect state due to their cache we have to
change the button status dynamically in JS.

Users with JS disabled will still see a button but it links to the talk to us
section where we detail the opening hours (and will eventually have an email
address for them to use).