Extend link checker to ensure internal ones work
Currently we run a nightly check of our external links to ensure they
resolve. Our internal link checker only runs as part of the CI process
where we don't talk to the API, so links to API-requiring pages like
events and the privacy policy are skipped.

This lead to a recent problem where we had a broken link that wasn't
picked up.

Here we're extending the link checker to do an internal run too.
Internal links are identified by starting with a `/`.