Swap order of page title so the name of the individual page comes first
The page title is shown in the browser's tab or as the window heading, and will be truncated if longer than the space available. 

Put the name of the individual page before `Get into Teaching` (instead of after) so that users can more easily see which page they are on when only the first part of the title is visible.