Hide covid banner by default
The covid banner is in a dismissed state for the Google Page Speed checker and
because we show it by default and hide it with Javascript we get a content
layout shift reported. To avoid this we can hide it by default and show it only
when it has not been dismissed.

Minor refactoring to the `BannerController` to use the new `*-value` syntax.

If Javascript is disabled we show the banner (which maintains the current
behaviour).