Prevent cookie banner from being the LCP element
Currently, if a user has not accepted our cookie policy the
banner text gets classified as the Largest Contentful Paint
element on pages without a hero image or a lot of text above the
fold (i.e. most event pages).

The only way I've found to get around this is to load the banner
visible but off-screen and then transition the banner to the
main viewport in JS if the user hasn't accepted cookies yet.