Remove CSP whitelist of Google domains
A number of Google domains (~200) have been whitelisted for Google Analytics across 4 directives (connect-src, script-src, img-src, style-src). 

However, it is recommended that headers are kept as small as possible. Remove the list, separately adding a Prometheus alert for any CSP violations that include `Google` so that the list can be built back up.