Improve back link text on cookie prefs page
When someone comes from an external referrer/unknown source or
directly onto the cookie preferences page we show 'Back to page'
even though we take them to the homepage and not where they came
from.

Instead, we can change this text to be 'Go to home page' for
better clarity and only show 'Back to page' if we actually know
the page they came from.