Fix possible error on search param
If a user manually enters `/search?search` then the `params.fetch` returns
`nil` and if they enter `/search?search=` then `params.fetch` will be an empty
string. We need to ensure that we don't call `permit` on either of these values
as it will result in an exception. Instead, we check the type is
`ActionController::Parameters` and, if not, we return `nil` to give an empty
result.