Remove closing </source> tags
When we perform image processing we manually modify the HTML response with
Nokogiri to inject `<source>` tags. By default Nokogiri will render closing
tags for empty elements, resulting in `<source></source>` which is not valid
HTML 5 (tags with no content should be self-closing).

It is possible to configure Nokogiri to self-close empty tags, however in doing
so it completely trashes other parts of the HTML document (we probably have
invalid HTML/an unclosed tag somewhere that's causing this). For now we can
manually remove the closing `</source>` tags.