Prevent users submitting the events search multiple times
When a user searches for a postcode that isn't in the cache, the API will look it up via Google Geocoding then persist it.

If the API gets multiple requests in quick succession, the requests will be processed concurrently, resulting in a duplicate key exception as the API tries to persist the same postcode multiple times (the key for the `Locations` table is the postcode).

We can prevent users double submitting the form with `data-disable-with` the message `Searching...` will be shown in the button while the request is processing.