Load Zendesk JS when user clicks chat button
Currently we load the Zendesk code on every page; even if we defer the loading
to the end it still has a significant impact on our page speed.

Instead, this commit doesn't load the Javascript until the user interacts with
the chat online button; at which point it will download the JS and then open
the chat window.

We could change the button state to be 'loading' while it downloads, but in my
testing its still reasonably responsive without it.