Add loading message when starting a new chat session
As we now download the JS snippet on-demand it could result in a short delay,
particularly on mobile/3G. To let the user know something is happening we want
to add a loading message during this process.

Subsequent clicks to start a chat on the same page do not download the JS, so
the loading message is only needed the first time.