Add Markdown linter and configure
Adds a Markdown linter and configures it so that it by-en-large passes on our
set of Markdown files.

The linter can be ran with:

```
mdl app/views/content/**/*.md
```

It is set to run during CI and fail if there are any linting errors present.
This may be too heavy-handed and blocking for content editors so we may end up
switching it off and running it nightly instead.