Replace mailing list completion page
The new content is less wordy and will provide a CTA that takes users to
the welcome guide.

As a stopgap, before the welcome guide is complete, I've created a
fallback where the old block (purple boxes) content is present in
production. This will give users a path to follow before the welcome
guide is ready to be released.