Add #prune! to Wizard::Store
The prune method is a little like purge except it allows some keys (none
by default) to be left in the app_data. This could be useful where we
want to display some information on the completion page (or beyond).

For now, if a user restarts the mailing list signup, the first page will
have the name they used on the previous run pre-filled. Data is all
stored in the session so assumed to be safe, although using a shared
machine could allow people to see previous users' data.

This would be the same had a user partially filled in the wizard then
stopped though, so we're not really introducing a new attack vector
here.