Remove and validate extra characters when creating event partial URL
Users can generate a partial URL based on the event name and start date of an event. Currently there is no validation, so we have seen events partial URLs with non-alphanumeric characters such as brackets, colons, and question marks.

Update the JavaScript that generates the partial URL to cut out any characters that are not alphanumeric or whitespace. 

Update the partial URL validation so that only alphanumeric, hyphens, and underscores are allowed.