Remove automated event closed message
We currently display an automated message on events that are 'closed'. Instead,
the events team want to use the existing `message` attribute to display a
message when the event is closed. We already render this when its filled on the
event page, so we simply need to remove the automated message.