Ensure extra fields remain in mailing list store
We want to leave some extra fields in the mailing list signup store so
we can personalise the welcome guide. The subject, journey stage and
degree status will be used in addition to the name to customise the
content that's displayed in the welcome guide.

To aid testing the store instance variable has been exposed as an
attr_reader