Add new age step to mailing list wizard
We want to display a new age step as part of the mailing list sign up so
that we can collect information about a candidates age.

Before we collect this data we want to evaluate three different methods for
collecting age-related information, in order to determine how the user
responds (does one method cause more drop-off during sign up than another).
The methods are:

- Date of birth
- Year of birth
- Age range

We are not currently persisting the data entered by the user.