Serve random (but consistent) age question
When the feature is enabled, we want to show the user a random age step in
the mailing list sign up. The step should be consistently shown in case the
user goes forward/backwards in the journey or starts over again.

I will make a ticket to come back to the `mailing_list_wizard_spec` test
cases as I think they could be cleaned up in general/made more concise.