Clean up the mailing list feature specs
These were fairly verbose and repetetive, which made the file
large and hard to read. In an effort to make them easier to
follow I've extracted common helpers (that could also be used for
the events sign up) into a `SpecHelper` and refactored the tests
to be more concise. The tests now also assert on the data being sent to the CRM, which should be more robust.