Support fallback email for when JS is disabled
When JS is disabled we need to show a fallback email address in the talk to us
section and change the title to 'Email us' instead of 'Chat to us'. When out of
support hours we want to hide the 'Email us' section entirely. This differs to
how we handle it for JS-enabled users (we retain the heading and replace the
button with 'Chat closed' test).

This is a temporary solution ahead of a wider redesign of the chat
component/CTAs; the code is complex and hard to follow due to all the different
ways we show a chat CTA and the variations for if chat is online/offline and JS
is enabled/disabled.