Handle skip verification requests
We will handle skip verification requests by having `?skip_verification`
appended to the authenticate step route. We can intercept this in the wizard
step controllers (via the concern) and enable the user to continue as
'unverified'.

A link has been added to the event sign up authenticate step to enable users to
leverage this feature (it is not available in other transactions).