Add new GTM container
We are going to migrate all of our tracking/analytics into a new GTM container
so that we can switch from the old one to the new one without any downtime.

Add the new container to the dev/rolling environments so we can test it locally
and in the review apps.