Inform GTM of consent changes
As we will be loading all of our JS tracking/analytics snippets in via GTM
going forward we need to be able to set the consent according to the user's
preferences.

We set the default consent to 'denied' with a delay of 500ms (allowing for the
async controller to update the consent). Once the controller connects it sends
the consent as stored in the preferences cookie and it also listens for any
changes made on the cookie preferences page, piping them through to GTM.