Shift GTM into own pack/class
Move the GTM setup into a class, removing the need for a controller and pull it
in as its own pack.

The consent wasn't reflecting properly, I think because we were defaulting it
all to denied then updating it with the correct consent later. Instead, having
it in its own pack we can pull in the cookie prefs and default it to the
current values straight-off before we load in any tags.