Add noscript pictures to end of parent
There's a slight display oddity in the footer where two images that are
next to each other in the markup don't have the correct spacing applied.

The spacing uses the css `picture + picture`, but there's a `<noscript>`
tag injected between them by the lazy loading images.

Now instead of adding the noscript right after the picture, we're adding
them all at the end of the parent, so instead of:

<picture> <noscript> <picture> <noscript>

it will be:

<picture> <picture> <noscript> <noscript>