Don't display provider information for TTT/QT events
We currently don't populate provider information in the CRM for Train to
Teach or Question Time events, however we _do_ want to start displaying this
information in the Google Structured Data for events. As we are going to
start populating the fields we need to add another conditional that hides
that information from the event details page in the website.