Fix tab behaviour with cookie dialog
The cookie dialog being within the page but hidden causes some issues with
tabbing behaviour on the page. When tabbing down the page it enters the
hidden cookie dialog elements before proceeding to the footer. In addition,
after accepting cookies the tabbing sequence remains within the dialog and
the rest of the page can't be tabbed through.

To fix this, the `tabindex` of the cookie dialog elements are initially set
to `-1`, which causes the browser to skip over them during regular tabbing.

When the cookie dialog opens we set the `tabindex` of elements to ensure the
first tab goes to the accept button. We listen for `keydown` events to ensure
the tab sequence goes through the dialog elements only whilst it is open.

When the dialog is closed we remove the `keydown` listeners to revert tabbing
back to the main page and focus/blur on the first anchor link to ensure the
tab sequence starts at the top of the page instead of the footer (which is
where the cookie acceptance HTML is located).

Lastly, we set the `tabindex` of the cookie dialog elements back to `-1` so
they are skipped over on any subsequent tabbing.