Remove ETag caching
Our ETag caching doesn't appear to currently work; the ETags don't match
between requests for some reason and even if I force them to match we never
serve a 304 Not Modified. I _think_ its because we have an explicit `render`
call in our `PagesController#show` action and according to the Rails docs the
not modified behaviour is invoked if we don't call `render`.

I'm removing the implementation for now (which means we will default to using
Rails ETags that compute on the response body rather than the last deployment
SHA - which means we should be no worse off anyway).

The next step is adding a static page cache, which should actually give us a
decent performance boost as we won't need to hit Rails after the first request
for a static page. It means ETags won't work but I expect this to be faster
regardless.