Fix image post-processing for hosted assets
When the app is running in the preprod/prod environment the assets
are hosted on a different domain, so the various image `src`
attributes contain absolute URLs instead of paths.

Update the `NextGenImages` and `ResponsiveImages` to account for
`src` attributes containing either a domain or path.

Even though the assets are hosted elsewhere they remain in the
docker image so the various `File` checks still work.