Correctly serve webp images
`Rack::Mime` doesn't support webp images out of the box; to get
around this we have to tell it to use the correct content type for
the extension.

Oddly, its been fine serving webp images as `text/plain` up until
we added in a `Access-Control-Allow-Origin` header when we started
serving assets on a different domain. A side-effect of this is that
the content-type has to now be correct or the request is blocked by
cross-origin read blocking (CORB). More info here:

https://www.chromium.org/Home/chromium-security/corb-for-developers

Specifically:

```
For example, a response served with a “X-Content-Type-Options: nosniff” response header and an incorrect “Content-Type” response header may be blocked. This could, for example, block an actual image which is mislabeled as “Content-Type: text/html” and “nosniff.” If this occurs and interferes with a page’s behavior, we recommend informing the website and requesting that they correct the “Content-Type” header for the response.
```