Flush asset cache
Updates the path prefix with a version in order to flush the assets
as we now have webp images cached in the staging CloudFront with
the wrong content-type (from my testing).

I can't find a better way of doing this; Webpack seems to let you
define filenames explcitly but I'd have to do it for every
extension type which is awkward. Sprockets lets you specify the
asset version which weighs into the digest value, however I can't
find a corresponding option for Webpack and instead the easiest
thing seems to be to change the pack path to include a version.