Fix multi-line highlight in Firefox
The multi-line text highlight was not rendering correctly in Firefox as of
Firefox 32 due to the the box-shadow method being used. FF 32 introduces a new
property: `box-decoration-break` that causes the breakage.

This property defaults to 'split' but needs to be specified as 'clone' to
achieve the desired multiline padding effect.