Perform beta redirect before static page cache
The beta redirect currently happens as a Rails route, which used to work fine
but now that the pages are statically cached the routes aren't evaluated on
every request. If the page matches a static page it will be served by Rack; to
re-instate the redirect we need to do so in Rack before the
`ActionDispatch::Static` middleware.