Update Access-Control-Allow-Origin header for test
We have multiple domains that point to our test
instance of the app; the
Access-Control-Allow-Origin header restricts
requests to our assets from certain domains,
however we can only specify a single domain in
this header which means fonts aren't loaded on
the other test domain.

A way around this is to change it to a wildcard
that accepts all requests; it should be no less
secure given we have basic auth on the test
instance anyway.

The only other solution would require a reverse
proxy such as NGINX that would let us set the
header dynamically if it comes from a list of
supported domains, but we don't have that
infrastructure at present.