Switch to loading Google Optimize on cookie acceptance
Initially I thought Google Optimize wouldn't drop a cookie until GA was pulled
in, however that doesn't appear to be the case so we need to prevent it firing
before the user has opted-in to non-functional cookies.

This creates a new `GoogleOptimizeController` that does that for the legacy
tracking pixel method; a follow-up commit will ensure it also works for the new
tracking pixel setup.