Send welcome guide variant on ML sign up
When a candidate signs up to the Mailing List we want to populate the
`welcome_guide_variant` attribute with the query string that can be used to
serve the correct version of the welcome guide in an email that will later be
sent to the candidate.

We only want to populate this field if the candidate qualifies for the welcome
guide, leaving it as `nil` if they don't.