Refactor global hidden style
There was a global `.hidden` class that was only applied when JS is enabled;
this is confusing and a pain to override, as you often want the element to be
hidden regardless of JS status when applying a `hidden` class.

Rename `.hidden` to `.hidden-when-js-enabled` and create a new `.hidden` class
that will always hide an element when applied. It's decorated with `important!`
so that custom components that have, for example, `display: flex` will still be
hidden when the class is applied.

Fix typo in style attribute that should be `margin-bottom`, not
`bottom-margin`.