Add ability to easily exclude a content page from sitemap/index
When we are running A/B tests we will need to de-index any variants
that we don't want to let Google know about.

Provide simple mechanism to exclude a page from the sitemap and serve
a noindex header; include `noindex: true` in the frontmatter.

Refactor our existing test A/B pages to use the new mechanism.