Return 404 for past teaching events
If a teaching event is in the past we no longer want to serve it to the user,
instead returning a 404 with the custom event not found page.

The 404 response should prompt Google to deindex the page and the not found
events page directs the user to search for another event.

We define "in the past" as being on the previous day, rather than a
time-granular comparison; this makes it consistent with the search results so
if we return an event the starts at 9am and its now 10am it will still be
accessible for the remainder of the day.

The exception to this is with Online Q&A events, which have an archive so that
users can still view transcripts/comments of the event after it has happened.
We want to still allow users to access these events and Google to index them.