Add new Google Optimize JS implementation
The current way we include Google Optimize works but is not GDPR compliant.
Before we can run experiments with users we need to ensure we are adhearing to
their cookie preferences. This, along with Turbolinks and only wanting to
include the script on an experiment page (due to the impact on performance)
makes the process much more complicated.

We achieve this by extracting the Google Optimize JS into its own pack thats
loaded synchronously (to avoid a page flicker if serving a variant). On the
initial page load we initialise the Optimize script if they have previously
consented and are on an experiment page. If they land on an experiment page and
have not yet accepted cookies we blur the background to avoid showing them a
variant. When they accept we reload the page, which will drop the Optimize
script as in the previous scenario. If a user transitions to an experiment page
via Turbolinks we force a full page load which drops the script (assuming they
have consented).