Switch noindex to meta tag method
We were preventing search engines from indexing a web page by returning a
`X-Robots-Tag` header. This works for the first request that hits Rails, but if
the page is then served by the static page cache middleware the header is no
longer present.

Instead, we can add a `meta` tag to the page, which will then be part of the
static page cache and return on every request.