Hide provider website URL for TTT/QT events
We didn't used to populate the provider website URL for TTT/QT events, however
since we are serving structured data we have asked the events team to fill it
in for all event types. We still want to hide the provider information for
TTT/QT events on our website though.