Trim white space before rendering Markdown placeholders
Some of the view components are wrapped in a conditional whereby the content is
then indented, for example:

```
<% if something %>
  <p>text</p>
<% end %>
```

Most of the time when this is injected into Markdown its absolutely fine,
however if it happens to be injected immediately after a bullet list the
leading spaces result in the component itself being included in the final
`<li></li>` content. For example:

```
* one
* two
$component$
```

Results in:

```
<ul>
  <li>one</li>
  <li>two</li>
  <li>[component-html]</li>
</ul>
```

To prevent this, we need to trim the leading white space prior to rendering
components so that they are never treated as a list item.