Add custom styling for the last page of blog posts
These are _usually_ a paragraph of information and not directly part of
the blog. Styling this way means we can allow them to stand out from the
rest of the post but still be maintained in the regular fashion inside
the blog files.

This also means writers aren't forced to italicise the final paragraph
for a bit of visual distinction.