Retry failed external link checks
We are seeing flakiness in the external link checker where it reports a
broken link and when we check it manually straight after, or perform a
re-run, the link check passes. To try and limit the impact of 'blips'
with external websites we can retry the external links multiple times.
The order is also shuffled which may help weed out flakiness.