Fix Google Optimize redirect experiments
Google Optimize will drop a cookie for 5 seconds after performing a
redirect; if that cookie is present it will not perform another redirect.
The idea is that this prevents redirect loops happening if you
incorrectly configure an experiment. The issue with it is that it means a
user can be redirected to a variant, then click the link again and end up
seeing the control if they do it quikly.

It doesn't feel great manually clearing this, but we're unlikely to be
running lots of experiments (which is where a redirect loop becomes
likely) and this ensures that users always see the correct variant every
time.