Fix accessibility issues with autocomplete input
We see a warning on Silktide that the input has no label. To
fix this we're going to try adding an aria-label to the
auto-complete input and also co-locate the label next to the
input. One or both of these may not be necessary, but we
won't know until the next Silktide report.