Add ability to track client-side metrics
We want to be able to track some metrics that trigger from the client-side. As
all our metrics go via Prometheus we will need to send the information to Rails
from the client in order to track it.

Add `ClientMetricsController` to accept incoming client metric information from
Javascript.

Add `app_client_cookie_consent_total` metric as the first client-side metric to
track the opt-in success rate.