No index event sign up page
We don't want users to land directly on this page; instead they should go
through the view event page.

Google Search Console also shows a redirect error here, as we redirect back to
the event page if the event is closed -- if the event also happens to be in the
past we will then return a 404 and I think Google has an issue with redirecting
to a 404 page.

Remove header method of no-indexing a page given that we specify this in the
HTML now (this should have been removed already but I must have missed it --
noindex via header doesn't work all the time because of our static page cache).