Refactor remaining constants into OptionSet class
We have a number of constants remaining in
`GetIntoTeachingApiClient::Constants` that we want to restructure to make
referencing them easier.

Pull out the constants into a new `OptionSet` class and add convenience
methods. I've lumped all these into one class to keep it concise and
extensible.

Arguably the hashes could be simplified with keys that map to more concise
symbols, but it feels worth keeping this matching the CRM values in case we
inadvertently expose them to the user. The only exception to this is the
mailing list channels which I've renamed slightly as they will never be
user-facing.