Rework the enter verification code step
This step doesn't perform all that well and we see roughly 1/5 users drop
off/don't successfully enter their verification code and progress to the next
step.

In an effort to tackle this we are making the copy clearer/reworking the layout
of the steThis step doesn't perform all that well and we see roughly 1/5 users
drop off/don't successfully enter their verification code and progress to the
next step.