Make image_meta_tags use asset_pack_url internally
Now we use a different host for the app and assets this method doubles
up the domain. Using asset_pack_url should resolve it.

Testing proved tricky, was unable to override the asset_path on a
per-test basis, so remove the domain from the existing test and add a
new one to ensure we're using asset_pack_url from within
image_meta_tags