Make welcome guide use path instead of query
We ran into a problem where Dynamics' link tracking appends ?p0= to the
links it generates making the subsquent query param be ignored. Here's
an example query:

?p0=?degree_status_id=222750001&preferred_teaching_subject_id=962655a1-2afa-e811-a981-000d3a276620
    ^
    |
    | from here is ignored

To combat it, this PR changes the format, we go from this:

/welcome?degree_status_id=222750001&preferred_teaching_subject_id=962655a1-2afa-e811-a981-000d3a276620

To this:

/welcome/email/subject/maths/degree-status/graduate

Two things have changed

* we're using the path instead of the query string (see the new routes)
* we're referring to records using a slug instead of a ID or GUID. This
  makes it easier and gives the user a better experience

In this app we convert the string name to 'slug' by lowercasing and
replacing spaces with underscores. The CRM doesn't know about them yet
so ideally we'd add this data (and hard code it) in the CRM and then use
that as the source.