Add send_event_emails attribute to postcode step
We want to show a yes/no toggle on the postcode step, so we only show the text
input when the user selects 'yes'.

Add necessary attribute to populate the form with.

Change postcode presence to be conditional based on if they select 'yes'.

Retain CRM postcode if they have one and select 'No' to ensure we
don't accidently clear it in the CRM.v