Pull in radio functionality from govuk-frontend
Pull in the radio button functionality from govuk-frontend in addition to the
error summaries. We only do s will allow the toggle to show field JS to work.
We only pull this into the registration pages; update the filename to be more
generic.

Move JS to pull in after the footer instead of the head tag; previously if you
submitted the form, went back (via Turbolinks page change) and toggled the
radio it would not show the field as the JS was not re-initialising.