Retry Net::ReadTimeout exceptions during test run
We occasionally see timeout errors on `features/chat_spec.rb`. The
source of thee flakiness isn't clear, so allowing retries for now to
hopefully avoid it failing builds outright.

The other alternative is to remove these specs as there is
reasonable coverage in the JS tests, however they don't cover the
scenario when JS is disabled (which these specs do).