Remove the welcome prefix from welcome path
Currently the email uses a mailmerge-like approach to create the
hyperlink, it uses the following format:

https://getintoteaching.education.gov.uk/welcome{{contact.dfe_welcomeguidestring}}

The previous attempt replaced it with the full path, including the
`/welcome`. This will cause people who've already had their string set
to query params `?preferred_teaching_subject_id=abc123` to land on the
homepage instead of the welcome guide.

Removing the `/welcome` prefix and handling it in the email is a better
option because it serves both the people who've already signed up
and those that will after this change has gone live.